+++ lib/rspec/expectations/handler.rb
@@ -46,8 +46,6 @@ module RSpec
     class PositiveExpectationHandler
       def self.handle_matcher(actual, initial_matcher, custom_message=nil, &block)
         ExpectationHelper.with_matcher(self, initial_matcher, custom_message) do |matcher|
-          return ::RSpec::Matchers::BuiltIn::PositiveOperatorMatcher.new(actual) unless initial_matcher
-
           match_result = matcher.matches?(actual, &block)
           if custom_message && match_result.respond_to?(:error_generator)
             match_result.error_generator.opts[:message] = custom_message
@@ -74,8 +72,6 @@ module RSpec
     class NegativeExpectationHandler
       def self.handle_matcher(actual, initial_matcher, custom_message=nil, &block)
         ExpectationHelper.with_matcher(self, initial_matcher, custom_message) do |matcher|
-          return ::RSpec::Matchers::BuiltIn::NegativeOperatorMatcher.new(actual) unless initial_matcher
-
           negated_match_result = does_not_match?(matcher, actual, &block)
           if custom_message && negated_match_result.respond_to?(:error_generator)
             negated_match_result.error_generator.opts[:message] = custom_message